Funpay-Telegram-Gifts/Funpay-Telegram-Gifts/requirements.txt
pyrofork>=2.3.68
tgcrypto>=1.2.5
python-dotenv>=1.0.1
colorlog>=6.8.2
qrcode
colorama>=0.4.6
rich>=13.7.0
psutil>=5.9.4
beautifulsoup4>=4.11.1
colorama>=0.4.6
requests==2.28.1
pytelegrambotapi==4.15.2
pillow>=9.3.0
requests_toolbelt==0.10.1
lxml>=5.3.0
bcrypt>=4.2.0